backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
langchain
langchain-openai
langchain-community
requests
pypdf
python-docx
python-multipart
tavily-python
ddgs
duckduckgo-search
chromadb>=0.4.0
langchain-chroma>=0.1.2
langchain-huggingface
sentence_transformers
zhipuai